fix(query): lower bare integers to xsd:integer in VALUES/inline data
SPARQL VALUES rows and JSON-LD inline values tagged a bare integer
literal xsd:long, while storage, the Turtle parser, arithmetic results,
and triple-pattern lowering all tag xsd:integer. The scan post-filter's
integer-family compatibility masked it on the join path, but in
term-identity contexts (DISTINCT, GROUP BY, sameTerm) and in result
serialization the same number surfaced as two distinct terms.

Lower a bare integer to xsd:integer at both boundaries
(fluree-db-sparql lower/term.rs VALUES rows; fluree-db-query
parse/lower.rs JSON-LD inline values), matching RDF 1.1 and storage
tagging. Lowering-time only. Scan-path datatype constraints and
dt_compatible are deliberately untouched (that is D5b, PR-X2).

No W3C register entry fails solely on this; regression coverage lives
in it_query_expression_semantics.rs (SPARQL result datatype, sameTerm
and DISTINCT collapse, plus the JSON-LD values-integer parity tests).

Fixes #1319.

+++ fluree-db-sparql/src/lower/term.rs
@@ -450,9 +450,14 @@ impl<E: IriEncoder> LoweringContext<'_, E> {
                     FlakeValue::String(value.to_string()),
                     lang.clone(),
                 )),
+                // RDF 1.1: a bare integer literal is xsd:integer. Storage,
+                // the Turtle parser, arithmetic results, and triple-pattern
+                // lowering all agree on xsd:integer — tagging VALUES rows
+                // xsd:long made the same number two distinct terms in
+                // term-identity contexts (DISTINCT/GROUP BY/sameTerm), #1319.
                 LiteralValue::Integer(i) => Ok(Binding::lit(
                     FlakeValue::Long(*i),
-                    Sid::new(XSD, xsd_names::LONG),
+                    Sid::new(XSD, xsd_names::INTEGER),
                 )),
                 LiteralValue::Double(d) => Ok(Binding::lit(
                     FlakeValue::Double(*d),
